Refactoring & small improvements
